make vertical tab bar
