+++ apps/dragonfly_browser/lib/src/screens/lobby/cubit/browser_interface_cubit.dart
@@ -28,7 +28,9 @@ class BrowserInterfaceState {
           RedockableInterface.searchBar,
           RedockableInterface.bookmarks,
         ],
-        leftDocks: [RedockableInterface.devtools],
+        leftDocks: [
+          RedockableInterface.tabBar,
+        ],
         bottomDocks: [],
         rightDocks: [
           RedockableInterface.devtools,
